requirements.txt
python-telegram-bot==13.7
requests==2.26.0
boto3==1.18.4